refactor domino and dspark code
