+++ configs/qwen3.5-4b-domino.json
@@ -1,11 +1,11 @@
 {
   "architectures": [
-    "DFlashDraftModel"
+    "DominoDraftModel"
   ],
   "attention_bias": false,
   "attention_dropout": 0.0,
   "auto_map": {
-    "AutoModel": "dflash.DFlashDraftModel"
+    "AutoModel": "domino.DominoDraftModel"
   },
   "block_size": 16,
   "bos_token_id": 248043,
