+++ specforge/modeling/draft/__init__.py
@@ -5,12 +5,16 @@ from .dflash import (
     extract_context_feature,
     sample,
 )
+from .domino import DominoDraftModel
+from .dspark import DSparkDraftModel
 from .llama3_eagle import LlamaForCausalLMEagle3
 from .registry import DRAFT_REGISTRY, available_drafts, register_draft, resolve_draft
 
 __all__ = [
     "Eagle3DraftModel",
     "DFlashDraftModel",
+    "DominoDraftModel",
+    "DSparkDraftModel",
     "LlamaForCausalLMEagle3",
     "build_target_layer_ids",
     "extract_context_feature",
