+++ tests/test_modeling/test_draft_registry.py
@@ -17,6 +17,8 @@ from specforge.modeling.auto import AutoDraftModelConfig, AutoEagle3DraftModel
 from specforge.modeling.draft import (
     DRAFT_REGISTRY,
     DFlashDraftModel,
+    DominoDraftModel,
+    DSparkDraftModel,
     LlamaForCausalLMEagle3,
     available_drafts,
     register_draft,
@@ -54,7 +56,7 @@ TINY_DFLASH = {
 
 TINY_DSPARK = {
     **TINY_DFLASH,
-    "architectures": ["DFlashDraftModel"],
+    "architectures": ["DSparkDraftModel"],
     "layer_types": ["full_attention"],
     "dflash_config": {
         "projector_type": "dspark",
@@ -63,6 +65,29 @@ TINY_DSPARK = {
         "confidence_head_alpha": 1.0,
         "confidence_head_with_markov": True,
     },
+}
+
+TINY_LEGACY_DSPARK = {
+    **TINY_DSPARK,
+    "architectures": ["DFlashDraftModel"],
+}
+
+TINY_DOMINO = {
+    **TINY_DFLASH,
+    "architectures": ["DominoDraftModel"],
+    "layer_types": ["full_attention"],
+    "dflash_config": {
+        "projector_type": "domino",
+        "emb_dim": 16,
+        "gru_hidden_dim": 16,
+        "pure_draft_prefix_len": 0,
+        "shift_label": False,
+    },
+}
+
+TINY_LEGACY_DOMINO = {
+    **TINY_DOMINO,
+    "architectures": ["DFlashDraftModel"],
 }
 
 
@@ -77,8 +102,12 @@ class DraftRegistryTest(unittest.TestCase):
     def test_builtin_architectures_registered(self):
         self.assertIn("LlamaForCausalLMEagle3", available_drafts())
         self.assertIn("DFlashDraftModel", available_drafts())
+        self.assertIn("DominoDraftModel", available_drafts())
+        self.assertIn("DSparkDraftModel", available_drafts())
         self.assertIs(resolve_draft("LlamaForCausalLMEagle3"), LlamaForCausalLMEagle3)
         self.assertIs(resolve_draft("DFlashDraftModel"), DFlashDraftModel)
+        self.assertIs(resolve_draft("DominoDraftModel"), DominoDraftModel)
+        self.assertIs(resolve_draft("DSparkDraftModel"), DSparkDraftModel)
 
     def test_unknown_architecture_raises_with_available_list(self):
         with self.assertRaises(KeyError) as ctx:
@@ -125,15 +154,67 @@ class AutoLoaderRegistryTest(unittest.TestCase):
         config = AutoDraftModelConfig.from_file(path)
         self.assertIsInstance(config, Qwen3Config)
 
-    def test_from_config_builds_dspark_as_dflash_projector(self):
+    def test_from_config_builds_dspark_as_explicit_draft(self):
         path = _write(TINY_DSPARK)
         self.addCleanup(os.unlink, path)
         config = AutoDraftModelConfig.from_file(path)
         model = AutoEagle3DraftModel.from_config(config)
+        self.assertIsInstance(model, DSparkDraftModel)
         self.assertIsInstance(model, DFlashDraftModel)
         self.assertEqual(model.projector_type, "dspark")
         self.assertIsNotNone(model.markov_head)
         self.assertIsNotNone(model.confidence_head)
+
+    def test_from_config_maps_legacy_dspark_projector_to_explicit_draft(self):
+        path = _write(TINY_LEGACY_DSPARK)
+        self.addCleanup(os.unlink, path)
+        config = AutoDraftModelConfig.from_file(path)
+        model = AutoEagle3DraftModel.from_config(config)
+        self.assertIsInstance(model, DSparkDraftModel)
+        self.assertIsInstance(model, DFlashDraftModel)
+        self.assertEqual(model.projector_type, "dspark")
+        self.assertIsNotNone(model.markov_head)
+
+    def test_from_config_maps_legacy_domino_projector_to_explicit_draft(self):
+        path = _write(TINY_LEGACY_DOMINO)
+        self.addCleanup(os.unlink, path)
+        config = AutoDraftModelConfig.from_file(path)
+        model = AutoEagle3DraftModel.from_config(config)
+        self.assertIsInstance(model, DominoDraftModel)
+        self.assertIsInstance(model, DFlashDraftModel)
+        self.assertEqual(model.projector_type, "domino")
+        self.assertIsNotNone(model.prefix_gru)
+
+    def test_from_config_builds_domino_as_explicit_draft(self):
+        path = _write(TINY_DOMINO)
+        self.addCleanup(os.unlink, path)
+        config = AutoDraftModelConfig.from_file(path)
+        model = AutoEagle3DraftModel.from_config(config)
+        self.assertIsInstance(model, DominoDraftModel)
+        self.assertIsInstance(model, DFlashDraftModel)
+        self.assertEqual(model.projector_type, "domino")
+        self.assertIsNotNone(model.prefix_gru)
+        self.assertIsNotNone(model.embed_proj)
+
+    def test_domino_uses_direct_state_dict_keys(self):
+        path = _write(TINY_DOMINO)
+        self.addCleanup(os.unlink, path)
+        config = AutoDraftModelConfig.from_file(path)
+        model = AutoEagle3DraftModel.from_config(config)
+        keys = set(model.state_dict())
+        self.assertTrue(any(key.startswith("prefix_gru.") for key in keys))
+        self.assertTrue(any(key.startswith("embed_proj.") for key in keys))
+        self.assertFalse(any(key.startswith("logit_head.") for key in keys))
+
+    def test_dspark_uses_direct_state_dict_keys(self):
+        path = _write(TINY_DSPARK)
+        self.addCleanup(os.unlink, path)
+        config = AutoDraftModelConfig.from_file(path)
+        model = AutoEagle3DraftModel.from_config(config)
+        keys = set(model.state_dict())
+        self.assertTrue(any(key.startswith("markov_head.") for key in keys))
+        self.assertTrue(any(key.startswith("confidence_head.") for key in keys))
+        self.assertFalse(any(key.startswith("logit_head.") for key in keys))
 
     def test_from_file_unknown_architecture_raises(self):
         path = _write({**TINY_EAGLE3, "architectures": ["NoSuchDraft"]})
